add window flags for drawing decorations in the future, also support xdg_toplevel <= 5

+++ wayland/util.odin
@@ -22,8 +22,8 @@ interface :: struct {
    events: [^]message,
 }
 array :: struct {
-   size: i64,
-   alloc: i64,
+   size: uint,
+   alloc: uint,
    data: rawptr,
 }
